parsed index.ts to create app.ts for test config
